Fixed a few MS VS warnings. JIRA CXX-7546
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@70077 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/services/netstorage_direct_nc.cpp
@@ -258,7 +258,7 @@ void SNetStorage_NetCacheBlob::SetExpiration(const CTimeout& ttl)
             ": infinite ttl for NetCache blobs is not implemented");
     }
 
-    m_NetCacheAPI.ProlongBlobLifetime(m_BlobKey, ttl.GetAsDouble());
+    m_NetCacheAPI.ProlongBlobLifetime(m_BlobKey, (unsigned)ttl.GetAsDouble());
 }
 
 void SNetStorage_NetCacheBlob::Close()
